refactor: changed NumberInput design by transfering from NumberIncrementDecrement
