+++ .changeset/honest-spies-dress.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+refactor: changed NumberInput design by transfering from NumberIncrementDecrement
